backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
geoalchemy2>=0.15.2
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
pydantic>=2.10.0
slowapi>=0.1.9
